implement dup2, get lseek to report proper seeking behavior
lseek would return a huge non-negative integer when 0 was meant to be returned
when setting the offset to 0 with SEEK_SET. This is because it was returning an off_t,
which doesn't fit in the MIPS 32-bit registers. To compromise, we don't allow files
larger than 16GB, as the offset is now a int32_t, which fits fine. It should really be a
uint32_t, but lots of old code in the system casts to off_t, which is signed 64 bit, so it's safe to cast 32-bit signed to 64-bit signed.

Add testbin/luketest to test open, close, lseek and dup and dup2 in the future. This is just a playground file for testing various things.

+++ kern/arch/mips/syscall/syscall.c
@@ -129,7 +129,7 @@ syscall(struct trapframe *tf)
 			err = sys_fstat((int)tf->tf_a0, (userptr_t)tf->tf_a1, &retval);
 			break;
 		case SYS_lseek:
-			err = sys_lseek((int)tf->tf_a0, (off_t)tf->tf_a1, (int)tf->tf_a2, &retval);
+			err = sys_lseek((int)tf->tf_a0, (int32_t)tf->tf_a1, (int)tf->tf_a2, &retval);
 			break;
 		case SYS_remove: /* called unlink in other unix-likes */
 			err = sys_remove((userptr_t)tf->tf_a0, &retval);
